Fix test-z3 link error: move tst_lcube into top-level src/test/lcube.cpp
The largest-cube unit test tst_lcube() lived in src/test/lp/lp.cpp, but
scripts/mk_make.py only compiles the top-level test directory (os.listdir,
non-recursive), so lp/lp.cpp is never linked into test-z3 in the Python make
build. Since main.cpp now registers X(lcube), the build failed with
'undefined reference to tst_lcube()' (CMake was unaffected because it lists
lp/lp.cpp explicitly).

Move the lcube_test namespace and tst_lcube() to a new src/test/lcube.cpp,
which both build systems compile, and add it to src/test/CMakeLists.txt.
Also make int_cube.h self-contained (include vector.h, numeric_pair.h,
lar_term.h) so it no longer depends on include order.
